Added storage and image handler
