+++ app/src/main/java/ch/epfl/sdp/peakar/gallery/ImageActivity.java
@@ -13,13 +13,13 @@ import android.view.WindowManager;
 import android.widget.ImageView;
 
 import ch.epfl.sdp.peakar.R;
+import ch.epfl.sdp.peakar.utils.ImageHandler;
 
 import java.io.IOException;
 
 public class ImageActivity extends AppCompatActivity {
 
     public static final String IMAGE_PATH_INTENT = "imagePath";
-    private static final int ROTATE_DEGREES = 90;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -40,51 +40,8 @@ public class ImageActivity extends AppCompatActivity {
      */
     private void setImage(String imagePath) {
         ImageView view = findViewById(R.id.fullscreen_image);
-        Bitmap imageBitmap = getBitmapUpwards(imagePath);
+        Bitmap imageBitmap = ImageHandler.getBitmapUpwards(imagePath);
 
         view.setImageBitmap(imageBitmap);
-    }
-
-    /**
-     * Get an image bitmap from a image path, where the image is the correct rotation
-     * @param imagePath path to image
-     * @return bitmap correctly rotated.
-     */
-    public static Bitmap getBitmapUpwards(String imagePath) {
-        int orientation = getOrientation(imagePath);
-        Bitmap imageBitmap = BitmapFactory.decodeFile(imagePath);
-        if (orientation == ExifInterface.ORIENTATION_ROTATE_90) {
-            imageBitmap = rotateBitmap(imageBitmap);
-        }
-        return imageBitmap;
-    }
-
-    /**
-     * Returns the orientation of an image in on an image path
-     * @param imagePath the path to image
-     * @return the orientation
-     */
-    private static int getOrientation(String imagePath) {
-        try {
-            ExifInterface exifInterface = new ExifInterface(imagePath);
-            return exifInterface.getAttributeInt(ExifInterface.TAG_ORIENTATION,
-                    ExifInterface.ORIENTATION_NORMAL);
-        }
-        catch (IOException e) {
-            e.printStackTrace();
-            return ExifInterface.ORIENTATION_NORMAL;
-        }
-    }
-
-    /**
-     * Rotate an image 90 degrees.
-     * @param imageBitmap the image to rotate
-     * @return the same image rotated 90 degrees
-     */
-    private static Bitmap rotateBitmap(Bitmap imageBitmap) {
-        Matrix matrix = new Matrix();
-        matrix.postRotate(ROTATE_DEGREES);
-        return Bitmap.createBitmap(imageBitmap, 0, 0,
-                imageBitmap.getWidth(), imageBitmap.getHeight(), matrix, true);
     }
 }
